Externalize Text animation logic into AnimatedText component
